requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
Authlib==0.11
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cvxopt==1.2.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dialogflow==0.4.0
dill==0.3.1.1
dlx==1.0.4
docloud==1.0.375
docplex==2.12.182
entrypoints==0.3
fastdtw==0.3.4
Flask==1.0.2
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
fuzzywuzzy==0.17.0
gast==0.2.2
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.22.0
h5py==2.9.0
httplib2==0.14.0
idna==2.6
inflection==0.3.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-polyfield==5.7
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==5.0.0
mpmath==1.1.0
mysql-connector==2.2.9
nbconvert==5.6.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
nest-asyncio==1.2.1
networkx==2.4
notebook==6.0.1
ntlm-auth==1.4.0
numpy==1.17.0
oauthlib==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psutil==5.6.7
ptyprocess==0.6.0
pusher==2.0.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pymongo==3.10.0
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyrsistent==0.15.4
pyscf==1.6.5
python-dateutil==2.8.0
python-dotenv==0.8.2
python-Levenshtein==0.12.0
pytz==2019.3
pyzmq==18.1.0
qiskit==0.14.0
qiskit-aer==0.3.4
qiskit-aqua==0.6.1
qiskit-ibmq-provider==0.4.4
qiskit-ignis==0.2.0
qiskit-terra==0.11.0
qtconsole==4.5.5
Quandl==3.4.8
requests==2.22.0
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.5
sympy==1.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.6.2
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wikipedia==1.4.0
wrapt==1.11.2
